requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
attrs=19.1.0=py36_1
backcall=0.1.0=py36_0
blas=1.0=mkl
bleach=3.1.0=py36_0
ca-certificates=2019.1.23=0
certifi=2019.3.9=py36_0
cffi=1.11.5=py36he75722e_1
cudatoolkit=8.0=3
cudnn=5.1=0
cycler=0.10.0=py36_0
dbus=1.13.6=h746ee38_0
decorator=4.4.0=py36_1
defusedxml=0.5.0=py36_1
docopt=0.6.2=py_1
entrypoints=0.3=py36_0
expat=2.2.6=he6710b0_0
faiss-gpu=1.4.0=py36_cuda8.0.61_1
fontconfig=2.13.0=h9420a91_0
freetype=2.9.1=h8a8886c_1
glib=2.56.2=hd408876_0
gmp=6.1.2=h6c8ec71_1
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=hb453b48_1
icu=58.2=h211956c_0
intel-openmp=2019.1=144
ipykernel=5.1.0=py36h39e3cac_0
ipython=6.0.0=py36_1
ipython_genutils=0.2.0=py36_0
jedi=0.13.3=py36_0
jinja2=2.10=py36_0
jpeg=9b=h024ee3a_2
jsonschema=3.0.1=py36_0
jupyter_client=5.2.4=py36_0
jupyter_core=4.4.0=py36_0
kiwisolver=1.0.1=py36hf484d3e_0
libedit=3.1.20181209=hc058e9b_0
libffi=3.2.1=hd88cf55_4
libgcc=7.2.0=h69d50b8_2
libgcc-ng=8.2.0=hdf63c60_1
libgfortran=3.0.0=1
libgfortran-ng=7.3.0=hdf63c60_0
libopenblas=0.2.20=h9ac9557_7
libpng=1.6.36=hbc83047_0
libsodium=1.0.16=h1bed415_0
libstdcxx-ng=8.2.0=hdf63c60_1
libtiff=4.0.10=h2733197_2
libtorch=0.1.12=nomkl_0
libuuid=1.0.3=h1bed415_2
libxcb=1.13=h1bed415_1
libxml2=2.9.9=he19cac6_0
llvm=3.3=0
markupsafe=1.1.1=py36h7b6447c_0
matplotlib=3.0.2=py36h5429711_0
mistune=0.8.4=py36h7b6447c_0
mkl=2019.1=144
mkl_fft=1.0.1=py36h3010b51_0
mkl_random=1.0.1=py36h629b387_0
nb_conda=2.2.1=py36_0
nb_conda_kernels=2.2.0=py36_1
nbconvert=5.4.1=py36_3
nbformat=4.4.0=py36h31c9010_0
nccl=1.3.4=cuda8.0_1
ncurses=6.1=he6710b0_1
networkx=2.2=py36_1
ninja=1.8.2=py36h6bb024c_1
nomkl=2.0=0
notebook=5.7.6=py36_0
numpy=1.14.2=py36_nomklh2b20989_1
numpy-base=1.14.3=py36h0ea5e3f_1
olefile=0.46=py36_0
openblas=0.2.19=0
openssl=1.1.1b=h7b6447c_1
pandas=0.24.1=py36he6710b0_0
pandoc=2.2.3.2=0
pandocfilters=1.4.2=py36_1
parso=0.3.4=py36_0
patsy=0.5.1=py36_0
pcre=8.42=h439df22_0
pexpect=4.6.0=py36_0
pickleshare=0.7.5=py36_0
pillow=5.4.1=py36h34e0f95_0
pip=19.0.1=py36_0
pipreqs=0.4.9=py_0
prometheus_client=0.6.0=py36_0
prompt_toolkit=1.0.15=py36h17d85b1_0
ptyprocess=0.6.0=py36_0
pycparser=2.19=py36_0
pygments=2.3.1=py36_0
pyparsing=2.3.1=py36_0
pyqt=5.9.2=py36h22d08a2_1
pyrsistent=0.14.11=py36h7b6447c_0
python=3.6.8=h0371630_0
python-dateutil=2.7.5=py36_0
pytorch=1.0.1=py3.6_cuda8.0.61_cudnn7.1.2_2
pytz=2018.9=py36_0
pyzmq=18.0.0=py36he6710b0_0
qt=5.9.7=h5867ecd_1
readline=7.0=h7b6447c_5
requests=2.13.0=py36_0
scipy=1.1.0=py36_nomklh9d22d0a_0
seaborn=0.9.0=py36_0
send2trash=1.5.0=py36_0
setuptools=40.8.0=py36_0
simplegeneric=0.8.1=py36_2
sip=4.19.13=py36he6710b0_0
six=1.12.0=py36_0
sqlite=3.26.0=h7b6447c_0
statsmodels=0.9.0=py36h035aef0_0
system=5.8=2
terminado=0.8.1=py36_1
testpath=0.4.2=py36_0
tk=8.6.8=hbc83047_0
torchvision=0.2.2=py_3
tornado=5.1.1=py36h7b6447c_0
traitlets=4.3.2=py36h674d592_0
wcwidth=0.1.7=py36hdf4376a_0
webencodings=0.5.1=py36_1
wheel=0.32.3=py36_0
xz=5.2.4=h14c3975_4
yarg=0.1.9=py_1
zeromq=4.3.1=he6710b0_3
zlib=1.2.11=h7b6447c_3
zstd=1.3.7=h0b5b093_0
